DINOserver/requirements.txt
torch
flask
timm
torchvision
tqdm
requests
scikit-learn
matplotlib
pillow
transformers
